+++ Assets/Sprites/MainMenu/menu_background.png.import
@@ -2,7 +2,7 @@
 
 importer="texture"
 type="CompressedTexture2D"
-uid="uid://dbr12kp0p51i1"
+uid="uid://bt2272yo8pm"
 path="res://.godot/imported/menu_background.png-229509a4ba6dbfa2480e49735800a30c.ctex"
 metadata={
 "vram_texture": false
