+++ Assets/Sprites/MainMenu/pause_background.png.import
@@ -2,7 +2,7 @@
 
 importer="texture"
 type="CompressedTexture2D"
-uid="uid://co0dn2wbsk5a"
+uid="uid://28h8t3c342gi"
 path="res://.godot/imported/pause_background.png-21865bb7935e608fbe1bf74ccdaaba8e.ctex"
 metadata={
 "vram_texture": false
