+++ Assets/TileSets/Blocks/Ore/tin_ore.png.import
@@ -2,7 +2,7 @@
 
 importer="texture"
 type="CompressedTexture2D"
-uid="uid://cvv2pmynx7q5x"
+uid="uid://dq1n8b4mc4wc0"
 path="res://.godot/imported/tin_ore.png-0cfdbe0c64b3fe48f6672653caa03785.ctex"
 metadata={
 "vram_texture": false
