+++ Assets/TileSets/Characters/Player/player.png.import
@@ -2,7 +2,7 @@
 
 importer="texture"
 type="CompressedTexture2D"
-uid="uid://dswqwmo8hiqq3"
+uid="uid://dfy7x21n0g8g5"
 path="res://.godot/imported/player.png-4752a419ac2aaadd2ca3c21b85bfb211.ctex"
 metadata={
 "vram_texture": false
