+++ Scenes/MainMenu/sound/master_sound.tscn
@@ -1,7 +1,7 @@
 [gd_scene load_steps=3 format=3 uid="uid://bjb5e0eaiu8pt"]
 
 [ext_resource type="FontFile" uid="uid://d2jrr5wn34h8l" path="res://Assets/Font/chava/Chava-Regular.ttf" id="1_t4vdv"]
-[ext_resource type="Script" path="res://Script/MainMenu/sound/MasterSound.cs" id="2_j2vga"]
+[ext_resource type="Script" path="res://Script/MainMenu/Sound/MasterSound.cs" id="2_j2vga"]
 
 [node name="master_sound" type="Control"]
 layout_mode = 3
