+++ Scenes/MainMenu/sound/sfx.tscn
@@ -1,7 +1,7 @@
 [gd_scene load_steps=3 format=3 uid="uid://cdchqbfo247kv"]
 
 [ext_resource type="FontFile" uid="uid://d2jrr5wn34h8l" path="res://Assets/Font/chava/Chava-Regular.ttf" id="1_chogj"]
-[ext_resource type="Script" path="res://Script/MainMenu/sound/Sfx.cs" id="2_ce5c0"]
+[ext_resource type="Script" path="res://Script/MainMenu/Sound/Sfx.cs" id="2_ce5c0"]
 
 [node name="sfx" type="Control"]
 layout_mode = 3
