+++ Scenes/World/world.tscn
@@ -2,19 +2,19 @@
 
 [ext_resource type="Script" path="res://Script/World/WorldNode.cs" id="1_3kq5q"]
 [ext_resource type="PackedScene" uid="uid://blugsjfr8pef1" path="res://Scenes/Player/player.tscn" id="1_sp1lt"]
-[ext_resource type="Texture2D" uid="uid://p6lcvcty6o8k" path="res://Assets/TileSets/Blocks/Ore/floor.png" id="2_73k5m"]
+[ext_resource type="Texture2D" uid="uid://b6eyo4skwc7ag" path="res://Assets/TileSets/Blocks/Ore/floor.png" id="2_73k5m"]
 [ext_resource type="PackedScene" uid="uid://duk2ea22qr0ev" path="res://Scenes/MainMenu/pause_menu.tscn" id="2_wgtww"]
-[ext_resource type="Texture2D" uid="uid://c4cyuqjd13mbb" path="res://Assets/TileSets/Blocks/Ore/stone.png" id="3_tjpen"]
-[ext_resource type="Texture2D" uid="uid://dcdor2efgle12" path="res://Assets/TileSets/Blocks/Ore/amethyst_ore.png" id="3_yrx3v"]
-[ext_resource type="Texture2D" uid="uid://dduvh8ejspytv" path="res://Assets/TileSets/Blocks/Ore/coal_ore.png" id="4_caj3c"]
-[ext_resource type="Texture2D" uid="uid://glnr4xt0ys83" path="res://Assets/TileSets/Blocks/Ore/iron_ore.png" id="5_7bgry"]
-[ext_resource type="Texture2D" uid="uid://caro2517be0pb" path="res://Assets/TileSets/Blocks/Ore/bedrock.png" id="5_ol8h2"]
-[ext_resource type="Texture2D" uid="uid://75xe1rav07cu" path="res://Assets/TileSets/Blocks/Ore/copper_ore.png" id="6_u7sen"]
-[ext_resource type="Texture2D" uid="uid://dcb1fec4gc4xd" path="res://Assets/TileSets/Blocks/Ore/tin_ore.png" id="7_5t01j"]
-[ext_resource type="Texture2D" uid="uid://bp2ckb8a0of8w" path="res://Assets/TileSets/Blocks/Ore/aluminum_ore.png" id="8_afogh"]
-[ext_resource type="Texture2D" uid="uid://cjea764myysvi" path="res://Assets/TileSets/Blocks/Ore/gold_ore.png" id="9_lq81w"]
-[ext_resource type="Texture2D" uid="uid://cyr3h5t376x70" path="res://Assets/TileSets/Blocks/Ore/diamond_ore.png" id="10_4k2vi"]
-[ext_resource type="Texture2D" uid="uid://dbngi2y874g7m" path="res://Assets/TileSets/Blocks/Ore/emerald_ore.png" id="11_302eb"]
+[ext_resource type="Texture2D" uid="uid://5d682scj8r8f" path="res://Assets/TileSets/Blocks/Ore/stone.png" id="3_tjpen"]
+[ext_resource type="Texture2D" uid="uid://r8vdfd0kdtwn" path="res://Assets/TileSets/Blocks/Ore/amethyst_ore.png" id="3_yrx3v"]
+[ext_resource type="Texture2D" uid="uid://btbgr4auyx7cm" path="res://Assets/TileSets/Blocks/Ore/coal_ore.png" id="4_caj3c"]
+[ext_resource type="Texture2D" uid="uid://cxf03myfu3o5b" path="res://Assets/TileSets/Blocks/Ore/iron_ore.png" id="5_7bgry"]
+[ext_resource type="Texture2D" uid="uid://345s61wi2xyb" path="res://Assets/TileSets/Blocks/Ore/bedrock.png" id="5_ol8h2"]
+[ext_resource type="Texture2D" uid="uid://bu0iyi3y6n4hd" path="res://Assets/TileSets/Blocks/Ore/copper_ore.png" id="6_u7sen"]
+[ext_resource type="Texture2D" uid="uid://cvv2pmynx7q5x" path="res://Assets/TileSets/Blocks/Ore/tin_ore.png" id="7_5t01j"]
+[ext_resource type="Texture2D" uid="uid://dkg4isn58eooh" path="res://Assets/TileSets/Blocks/Ore/aluminum_ore.png" id="8_afogh"]
+[ext_resource type="Texture2D" uid="uid://ds3jkf76768o2" path="res://Assets/TileSets/Blocks/Ore/gold_ore.png" id="9_lq81w"]
+[ext_resource type="Texture2D" uid="uid://djmpag84504ii" path="res://Assets/TileSets/Blocks/Ore/diamond_ore.png" id="10_4k2vi"]
+[ext_resource type="Texture2D" uid="uid://dmdkkf5feae1k" path="res://Assets/TileSets/Blocks/Ore/emerald_ore.png" id="11_302eb"]
 
 [sub_resource type="TileSetAtlasSource" id="TileSetAtlasSource_cjwib"]
 texture = ExtResource("2_73k5m")
