+++ Script/MainMenu/graphics/vsync_mode.gd
@@ -4,6 +4,6 @@ extends Control
 
 func _on_check_button_toggled(toggled_on):
 	if toggled_on == true:
-		DisplayServer.VSYNC_ENABLED
+		DisplayServer.window_set_vsync_mode(DisplayServer.VSYNC_ENABLED)
 	else:
-		DisplayServer.VSYNC_DISABLED
+		DisplayServer.window_set_vsync_mode(DisplayServer.VSYNC_DISABLED)
